refactor(ui): change file name
