--- packages/ui/src/design-system/check-box/index.ts
@@ -1,1 +1,0 @@
-export { Checkbox } from './check-box.component';
