+++ packages/ui/src/design-system/checkbox/index.ts
@@ -1,0 +1,1 @@
+export { Checkbox } from './checkbox.component';
